Make ShardedTensor ctor more inline with torch.Tensor ctor
torch.Tensor ctor creates an empty tensor and this PR makes
ShardedTensor on par with that.

In particular we remove TensorInitParams and instead always a create an empty
tensor and then fill it in for things like ones, zeros, full etc. This is
inline with torch.ones etc. as well since even for those APIs we first create
an empty tensor and then fill it out.

Differential Revision: [D33934603](https://our.internmc.facebook.com/intern/diff/D33934603/)

ghstack-source-id: 148167263
Pull Request resolved: https://github.com/pytorch/pytorch/pull/72164